More changes
Added comma separators to MP, Physical/Magic Attack, Physical/Magic Defense,
Accuracy, and EXP. Also fixed a mistake with time-limited stats

+++ WzComparerR2.Common/CharaSim/Mob.cs
@@ -28,7 +28,7 @@ namespace WzComparerR2.CharaSim
         public string DefaultMP { get; set; }
         public string FinalMaxHP { get; set; }
         public string FinalMaxMP { get; set; }
-        public int MaxHP { get; set; }
+        public long MaxHP { get; set; }
         public int MaxMP { get; set; }
         public int HPRecovery { get; set; }
         public int MPRecovery { get; set; }
@@ -91,7 +91,7 @@ namespace WzComparerR2.CharaSim
                         case "defaultMP": mobInfo.DefaultMP = propNode.GetValueEx<string>(null); break;
                         case "finalmaxHP": mobInfo.FinalMaxHP = propNode.GetValueEx<string>(null); break;
                         case "finalmaxMP": mobInfo.FinalMaxMP = propNode.GetValueEx<string>(null); break;
-                        case "maxHP": mobInfo.MaxHP = propNode.GetValueEx<int>(0); break;
+                        case "maxHP": mobInfo.MaxHP = propNode.GetValueEx<long>(0); break;
                         case "maxMP": mobInfo.MaxMP = propNode.GetValueEx<int>(0); break;
                         case "hpRecovery": mobInfo.HPRecovery = propNode.GetValueEx<int>(0); break;
                         case "mpRecovery": mobInfo.MPRecovery = propNode.GetValueEx<int>(0); break;
